Save snaps generated during image praparation

+++ tests/lib/nested.sh
@@ -520,7 +520,9 @@ nested_prepare_snapd() {
             snap_id="PMrrV4ml8uWuEUDBT8dSGnKUYbevVhc4"
         fi
 
-        "$TESTSTOOLS"/snaps-state repack_snapd_deb_into_snap "$snap_name" "$NESTED_ASSETS_DIR"
+        if [ ! -f "$NESTED_ASSETS_DIR/$output_name" ]; then
+            "$TESTSTOOLS"/snaps-state repack_snapd_deb_into_snap "$snap_name" "$NESTED_ASSETS_DIR"
+        fi
         cp "$NESTED_ASSETS_DIR/$output_name" "$(nested_get_extra_snaps_path)/$output_name"
 
         # sign the snapd snap with fakestore if requested
@@ -539,28 +541,33 @@ nested_prepare_kernel() {
         snap_id="pYVQrBcKmBa0mZ4CCN7ExT6jH8rY1hza"
         version="$(nested_get_version)"
 
-        if nested_is_core_16_system || nested_is_core_18_system; then
-            kernel_snap=pc-kernel-new.snap
-            repack_kernel_snap "$kernel_snap"
-        elif nested_is_core_20_system || nested_is_core_22_system; then
-            snap download --basename=pc-kernel --channel="$version/edge" pc-kernel
+        if [ ! -f "$NESTED_ASSETS_DIR/$output_name" ]; then
+            if nested_is_core_16_system || nested_is_core_18_system; then
+                kernel_snap=pc-kernel-new.snap
+                repack_kernel_snap "$kernel_snap"
 
-            # set the unix bump time if the NESTED_* var is set,
-            # otherwise leave it empty
-            local epochBumpTime
-            epochBumpTime=${NESTED_CORE20_INITRAMFS_EPOCH_TIMESTAMP:-}
-            if [ -n "$epochBumpTime" ]; then
-                epochBumpTime="--epoch-bump-time=$epochBumpTime"
+            elif nested_is_core_20_system || nested_is_core_22_system; then
+                snap download --basename=pc-kernel --channel="$version/edge" pc-kernel
+
+                # set the unix bump time if the NESTED_* var is set,
+                # otherwise leave it empty
+                local epochBumpTime
+                epochBumpTime=${NESTED_CORE20_INITRAMFS_EPOCH_TIMESTAMP:-}
+                if [ -n "$epochBumpTime" ]; then
+                    epochBumpTime="--epoch-bump-time=$epochBumpTime"
+                fi
+
+                uc20_build_initramfs_kernel_snap "$PWD/pc-kernel.snap" "$NESTED_ASSETS_DIR" "$epochBumpTime"
+                rm -f "$PWD/pc-kernel.snap"
+
+                # Prepare the pc kernel snap
+                kernel_snap=$(ls "$NESTED_ASSETS_DIR"/pc-kernel_*.snap)
+                chmod 0600 "$kernel_snap"
             fi
-
-            uc20_build_initramfs_kernel_snap "$PWD/pc-kernel.snap" "$NESTED_ASSETS_DIR" "$epochBumpTime"
-            rm -f "$PWD/pc-kernel.snap"
-
-            # Prepare the pc kernel snap
-            kernel_snap=$(ls "$NESTED_ASSETS_DIR"/pc-kernel_*.snap)
-            chmod 0600 "$kernel_snap"
+            cp "$kernel_snap" "$NESTED_ASSETS_DIR/$output_name"
         fi
-        cp "$kernel_snap" "$(nested_get_extra_snaps_path)/$output_name"
+        cp "$NESTED_ASSETS_DIR/$output_name" "$(nested_get_extra_snaps_path)/$output_name"
+
         # sign the pc-kernel snap with fakestore if requested
         if [ "$NESTED_SIGN_SNAPS_FAKESTORE" = "true" ]; then
             make_snap_installable_with_id --noack "$NESTED_FAKESTORE_BLOB_DIR" "$(nested_get_extra_snaps_path)/$output_name" "$snap_id"
@@ -584,6 +591,8 @@ nested_prepare_gadget() {
                 fi
                 return
             fi
+
+            if [ ! -f "$NESTED_ASSETS_DIR/$output_name" ]; then
 
             # XXX: deal with [ "$NESTED_ENABLE_SECURE_BOOT" != "true" ] && [ "$NESTED_ENABLE_TPM" != "true" ]
             echo "Repacking pc snap"
@@ -658,6 +667,7 @@ nested_prepare_base() {
             snap_name="core22"
             snap_id="amcUKQILKXHHTlmSa7NMdnXSx02dNeeT"
         fi
+        output_name="${snap_name}.snap"
 
         existing_snap=$(find "$(nested_get_extra_snaps_path)" -name "${snap_name}*.snap")
         if [ -n "$existing_snap" ]; then
@@ -668,11 +678,15 @@ nested_prepare_base() {
             return
         fi
         
-        echo "Repacking $snap_name snap"
-        snap download --channel="$CORE_CHANNEL" --basename="$snap_name" "$snap_name"
-        repack_core_snap_with_tweaks "${snap_name}.snap" "new-${snap_name}.snap"
+        if [ ! -f "$NESTED_ASSETS_DIR/$output_name" ]; then
+            echo "Repacking $snap_name snap"
+            snap download --channel="$CORE_CHANNEL" --basename="$snap_name" "$snap_name"
+            repack_core_snap_with_tweaks "${snap_name}.snap" "new-${snap_name}.snap"
 
-        cp "new-${snap_name}.snap" "$(nested_get_extra_snaps_path)/${snap_name}.snap"
+            cp "new-${snap_name}.snap" "$NESTED_ASSETS_DIR/$output_name"
+        fi
+        cp "$NESTED_ASSETS_DIR/$output_name" "$(nested_get_extra_snaps_path)/$output_name"
+
         # sign the base snap with fakestore if requested
         if [ "$NESTED_SIGN_SNAPS_FAKESTORE" = "true" ]; then
             make_snap_installable_with_id --noack "$NESTED_FAKESTORE_BLOB_DIR" "$(nested_get_extra_snaps_path)/${snap_name}.snap" "$snap_id"
@@ -961,8 +975,8 @@ nested_start_core_vm_unit() {
     # the caller can override PARAM_MEM
     local PARAM_MEM PARAM_SMP
     if [ "$SPREAD_BACKEND" = "google-nested" ]; then
-        PARAM_MEM="${NESTED_PARAM_MEM:--m 4096}"
-        PARAM_SMP="-smp 2"
+        PARAM_MEM="${NESTED_PARAM_MEM:--m 8182}"
+        PARAM_SMP="-smp 4"
     elif [ "$SPREAD_BACKEND" = "qemu-nested" ]; then
         PARAM_MEM="${NESTED_PARAM_MEM:--m 2048}"
         PARAM_SMP="-smp 1"
